requirements.txt
Flask==3.0.0
google-cloud-datastore==2.18.0
python-jose
flask-cors
six
python-dotenv
requests
authlib
protobuf==3.20.*
